requirements.txt
requests>=2.31.0
PyYAML>=6.0
matplotlib>=3.7.0
gTTS>=2.3.0
